Fix attempt to use an obsolete SDL unicode event property.
This worked with a prerelease version of SDL 2.0 but broke in the
released version.

This also rips out the commented-out-anyway x11 paste support sketches
that were in the same code block. If anyone needs it, it's in
the revision history.

Note that this leaves support for only lowercase a-z.
TODO: use the real SDL2 Unicode text input events to fix this.