Fix text input.
This uses the SDL_TEXTINPUT event, new in SDL 2.0, to capture text input.
Due to the limitations of the current event_t type, it doesn't handle
fancy multicharacter input methods. (And I'm not positive that
gui_textinput.cc always does the right thing with the utf16 characters
that it produces.)